Create integration test for University to ensure API calling and parsing is working as intended
